update project handlers for calendar
