+++ apps/crn-server/src/config.ts
@@ -32,6 +32,7 @@ const {
   EVENT_SOURCE,
   GOOGLE_API_CREDENTIALS_SECRET_ID,
   GOOGLE_API_TOKEN,
+  GOOGLE_CALENDER_EVENT_QUEUE_URL,
   LOG_ENABLED,
   LOG_LEVEL,
   NODE_ENV,
@@ -81,6 +82,8 @@ export const googleApiCredentialsSecretId =
   GOOGLE_API_CREDENTIALS_SECRET_ID || 'google-api-credentials-dev';
 export const googleApiToken = GOOGLE_API_TOKEN || 'asap-google-api-token';
 export const googleApiUrl = 'https://www.googleapis.com/';
+export const googleCalenderEventQueueUrl =
+  GOOGLE_CALENDER_EVENT_QUEUE_URL || 'google-calendar-event-queue-url';
 export const isContentfulEnabled = true;
 export const logEnabled = NODE_ENV === 'production' || LOG_ENABLED === 'true';
 export const logLevel = LOG_LEVEL || 'info';
