+++ apps/gp2-server/src/handlers/webhooks/events-updated.ts
@@ -1,43 +1,17 @@
 /* istanbul ignore file */
 import { Handler } from 'aws-lambda';
 
-import {
-  getJWTCredentialsFactory,
-  syncCalendarFactory,
-  syncEventFactory,
-  webhookEventUpdatedHandlerFactory,
-} from '@asap-hub/server-common';
-import {
-  googleApiCredentialsSecretId,
-  googleApiToken,
-  region,
-} from '../../config';
-import Events from '../../controllers/event.controller';
-import { getCalendarDataProvider } from '../../dependencies/calendar.dependency';
-import { getEventDataProvider } from '../../dependencies/event.dependency';
-import { getContentfulGraphQLClientFactory } from '../../dependencies/clients.dependency';
+import { webhookEventUpdatedHandlerFactory } from '@asap-hub/server-common';
+import { SQSClient } from '@aws-sdk/client-sqs';
+import { googleApiToken, googleCalenderEventQueueUrl } from '../../config';
 import logger from '../../utils/logger';
 import { sentryWrapper } from '../../utils/sentry-wrapper';
 
-const getJWTCredentials = getJWTCredentialsFactory({
-  googleApiCredentialsSecretId,
-  region,
-});
-
-const graphQLClient = getContentfulGraphQLClientFactory();
-
-const eventController = new Events(getEventDataProvider(graphQLClient));
-const syncCalendar = syncCalendarFactory(
-  syncEventFactory(eventController, logger),
-  getJWTCredentials,
-  logger,
-);
-
+const sqs = new SQSClient();
 export const handler: Handler = sentryWrapper(
   webhookEventUpdatedHandlerFactory(
-    getCalendarDataProvider(graphQLClient),
-    syncCalendar,
+    sqs,
+    { googleApiToken, googleCalenderEventQueueUrl },
     logger,
-    { googleApiToken },
   ),
 );
