+++ packages/server-common/src/handlers/webhooks/index.ts
@@ -1,4 +1,5 @@
 export * from './connect-by-code.handler';
 export * from './contentful-poller.handler';
 export * from './contentful.handler';
+export * from './events-updated-process.handler';
 export * from './events-updated.handler';
